+++ RELEASES.md
@@ -11,6 +11,10 @@ Verison 1.7.0 (unreleased)
   * Arithmetic operations with one `Integer` operand and one
     `Rational` operand were added.
   * The method `Integer::div_exact_from` was added.
+  * The methods `Integer::gcd_u`, `Integer::gcd_u_mut` and
+    `Integer::gcd_u_ref` were added.
+  * The methods `Integer::lcm_u`, `Integer::lcm_u_mut` and
+    `Integer::lcm_u_ref` were added.
 
 Verison 1.6.0 (2019-09-03)
 ==========================